fixed arithmetic cost computation

+++ costmodel.py
@@ -7,6 +7,6 @@ def cost(mem, arith, tot_lines):
     #return cost
     #mc = mem['memops']/tot_lines
     mc = 1
-    ac = (arith['add']+arith['sub']+8*arith['mul']+10*arith['div']+10*arith['mod'])/tot_lines    #weighted arithmetic cost per line
+    ac = float(arith['add']+arith['sub']+8*arith['mul']+10*arith['div']+10*arith['mod'])/float(tot_lines)    #weighted arithmetic cost per line
 
     return {'memcost':mc, 'arithcost': ac}
